small fixes

+++ resources/shaders/Passthrough.vert
@@ -1,7 +1,7 @@
 #version 450 core
 
 // Input vertex data, different for all executions of this shader.
-layout(location = 0) in ivec3 vertexPosition_modelspace;
+layout(location = 0) in vec3 vertexPosition_modelspace;
 
 // Output data ; will be interpolated for each fragment.
 out vec2 UV;
